Phase 3: Breakup Timeline Page
